requirements.txt
Django==2.2.12
django-model-utils==4.0.0
python-dateutil==2.8.1